Beat map + segment-play in label editor; extract BeatPanel for reuse
Two UX improvements for /label/editor, plus a refactor.

## Beat map in the editor

Drops BeatPanel under the video so users can see whether a labeled
window aligns with the music's beat and phrase structure. Same ticks
/ downbeats / phrase markers / off-beat dots / playhead as the
analysis page. Click on the strip seeks the video.

## Click a pattern block → play just that segment

Each AI-predicted block + each existing user label is now clickable.
Clicking seeks the video to the pattern's start_time, plays it, and
auto-pauses when currentTime reaches the end_time. Lets the user
isolate a 3-second window to decide "is this a whip or a side pass?"
without scrubbing manually.

Implementation: videoRef + segmentEnd state + a useEffect that
watches currentTime and pauses on boundary. Play/Edit/Accept buttons
stop propagation so they still fire without triggering the row's
play handler.

## Refactor: BeatPanel moved to its own file

Was inline inside timeline-view.tsx (~220 lines). Moved to
`src/components/analyze/beat-panel.tsx` so the label editor and any
future callers can import it without pulling in the whole
TimelineView component. timeline-view.tsx now imports from there;
zero behavior change on the analysis page.

## Files

- new `src/components/analyze/beat-panel.tsx` — BeatPanel + BeatStrip
- `src/components/analyze/timeline-view.tsx` — replace inline defs
  with the import (-219 lines)
- `src/app/(app)/label/editor/page.tsx` — BeatPanel render, videoRef,
  offBeats parser, playSegment handler, click-to-play rows

## Verified

- `npx tsc --noEmit` clean
- Click-stopPropagation on action button groups so Accept / Fix /
  Edit / Delete still work independently of the row click
- segmentEnd clears after fire (one-shot) so the next manual play
  doesn't get auto-paused
